requirements.txt
pyspark>=3.5.0
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.18.0
scikit-learn>=1.3.0
pyyaml>=6.0
python-dotenv>=1.0.0
jupyter>=1.0.0
notebook>=7.0.0
